Alter Search behavior to update when project and date inputs change.
This change causes the Summary view to update when:

* The Project dropdown changes
* The start date changes
* The end date changes

This allows users to change the summary view without the extra Search
button click.

This change also alters how the related HTML elements are discovered in
the DOM. Each element is queried before event execution, which means
they do not need to be queried every time the Search events fire.